ios fixes
