+++ plugin.xml
@@ -71,39 +71,55 @@
 
     </platform>
 	
- 	<!-- IOS CONFIGURATION -->
+ <!-- IOS CONFIGURATION -->
     <platform name="ios">
         <config-file target="config.xml" parent="/*">
             <feature name="FCMPlugin">
                 <param name="ios-package" value="FCMPlugin"/>
             </feature>
         </config-file>
-
-        <header-file src="src/ios/FCMPlugin.h"/>
-        <source-file src="src/ios/FCMPlugin.m"/>
-        <header-file src="src/ios/AppDelegate+FCMPlugin.h"/>
-        <source-file src="src/ios/AppDelegate+FCMPlugin.m"/>
-        <resource-file src="../../GoogleService-Info.plist" target="Resources/GoogleService-Info.plist" />
-
-        <!-- FIREBASE LIBS -->
-        <header-file src="src/ios/firebase/Firebase.h"/>
-        <framework src="src/ios/firebase/FirebaseAnalytics.framework" custom="true"/>
-        <framework src="src/ios/firebase/FirebaseInstanceID.framework" custom="true"/>
-        <framework src="src/ios/firebase/FirebaseMessaging.framework" custom="true"/>
-        <framework src="src/ios/firebase/GoogleInterchangeUtilities.framework" custom="true"/>
-        <framework src="src/ios/firebase/GoogleIPhoneUtilities.framework" custom="true"/>
-        <framework src="src/ios/firebase/GoogleSymbolUtilities.framework" custom="true"/>
-        <framework src="src/ios/firebase/GoogleUtilities.framework" custom="true"/>
-
-        <!-- OTHER NEEDED LIBS -->
-        <framework src="libc++.tbd"/>
-        <framework src="libz.tbd"/>
-        <framework src="AddressBook.framework"/>
-        <framework src="libsqlite3.0.tbd"/>
-        <framework src="libz.tbd"/>
-        <framework src="Security.framework"/>
+		
+		<!-- ENABLE PUSH CAPABILITY -->
+		<config-file parent="aps-environment" target="*/Entitlements-Debug.plist">
+		    <string>development</string>
+		</config-file>
+		<config-file parent="aps-environment" target="*/Entitlements-Release.plist">
+			<string>production</string>
+		</config-file>
+		
+		<resource-file src="src/ios/GoogleService-Info.plist" target="Resources/GoogleService-Info.plist" />
+		
+        <header-file src="src/ios/FCMPlugin.h" />
+        <source-file src="src/ios/FCMPlugin.m" />
+		
+		<header-file src="src/ios/AppDelegate+FCMPlugin.h" />
+		<source-file src="src/ios/AppDelegate+FCMPlugin.m" />
+		
+		<!-- FIREBASE COCOA PODS (not for the moment)-->
+		<!-- <framework src="Firebase/Messaging" type="podspec" spec="~> 0"/> -->
+		
+		<!-- FIREBASE LIBS -->
+		<header-file src="src/ios/firebase/Firebase.h" />
+		<framework src="src/ios/firebase/FirebaseAnalytics.framework" custom="true" />
+		<framework src="src/ios/firebase/FirebaseCore.framework" custom="true" />
+		<framework src="src/ios/firebase/FirebaseInstanceID.framework" custom="true" />
+		<framework src="src/ios/firebase/FirebaseMessaging.framework" custom="true" />
+		<framework src="src/ios/firebase/GoogleInterchangeUtilities.framework" custom="true" />
+		<framework src="src/ios/firebase/GoogleIPhoneUtilities.framework" custom="true" />
+		<framework src="src/ios/firebase/GoogleSymbolUtilities.framework" custom="true" />
+		<framework src="src/ios/firebase/GoogleUtilities.framework" custom="true" />
+			
+		<!-- OTHER NEEDED LIBS -->
+		<framework src="libc++.tbd" />
+		<framework src="libz.tbd" />
+		<framework src="AddressBook.framework" />
+		<framework src="libsqlite3.0.tbd" />
+		<framework src="libz.tbd" />
+		<framework src="Security.framework" />
     </platform>
 	
-    <hook src="scripts/after_prepare.js" type="after_prepare"/>
+	<!-- COPY FCM CONFIG FILES TO PLATFORM AFTER PLUGIN INSTALL OR PLATFORM PREPARE -->
+	<hook src="scripts/fcm_config_files_process.js" type="after_plugin_add" />
+	<hook src="scripts/fcm_config_files_process.js" type="after_prepare" />
 	
 </plugin>
