+++ src/ios/firebase/FirebaseAnalytics.framework/Modules/module.modulemap
@@ -15,10 +15,7 @@ framework module FirebaseAnalytics {
   header "FIRUserPropertyNames.h"
 
   link framework "AddressBook"
-  link framework "AdSupport"
-  link framework "SafariServices"
   link framework "StoreKit"
-  link framework "SystemConfiguration"
 
   link "c++"
   link "sqlite3"
